Stand-alone options page; Streamline Help
* Move to a stand-alone options page on all browsers.
* Streamline the content of the Help page.
* Streamline styles.
