+++ package.json
@@ -11,9 +11,10 @@
     "profile-guarding": "node scripts/profile-guarding.js",
     "start:firefox": "web-ext run --browser-console --source-dir build/firefox/",
     "pretest": "eslint .",
-    "test": "npm run test:landmarks && npm run test:contrast && npm run test:migration-manager",
+    "test": "npm run test:landmarks && npm run test:contrast && npm run test:migration-manager && npm run test:css",
     "pretest:contrast": "rollup src/code/contrastChecker.js -f cjs > test/test-code-in-harness-contrast.js",
     "test:contrast": "node test/test-contrast.js",
+    "test:css": "npx stylelint --config .stylelintrc.json src/**/*.css",
     "pretest:landmarks": "rollup src/code/landmarksFinder.js -f cjs > test/test-code-in-harness-landmarks.js",
     "test:landmarks": "node test/test-landmarks.js",
     "pretest:migration-manager": "rollup src/code/migrationManager.js -f cjs > test/test-code-in-harness-migration-manager.js",
@@ -42,6 +43,8 @@
     "rollup-plugin-esformatter": "~0.7",
     "rollup-plugin-terser": "~4.0",
     "stats-lite": "~2.2",
+    "stylelint": "^13.0.0",
+    "stylelint-config-standard": "^19.0.0",
     "test": "~0.6",
     "yargs": "~13.2"
   }
