+++ src/assemble/manifest.chrome.json
@@ -13,7 +13,7 @@
   },
 
   "options_ui": {
-    "chrome_style": true
+    "chrome_style": false
   },
 
   "permissions": ["<all_urls>"]
