+++ src/assemble/manifest.common.json
@@ -57,7 +57,8 @@
   },
 
   "options_ui": {
-    "page": "options.html"
+    "page": "options.html",
+    "open_in_tab": true
   },
 
   "browser_action": {
