+++ src/assemble/manifest.firefox.json
@@ -24,7 +24,7 @@
   },
 
   "options_ui": {
-    "browser_style": true
+    "browser_style": false
   },
 
   "applications": {
