+++ src/assemble/messages.common.en_GB.json
@@ -69,7 +69,7 @@
   },
 
   "prefsBorderType": {
-    "message": "Border type:"
+    "message": "Border type"
   },
 
   "prefsMomentary": {
